requirements.txt
asgiref==3.3.1
astroid==2.5.2
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
Django==3.1.7
django-crispy-forms==1.11.2
django-filter==2.4.0
djangorestframework==3.12.4
future==0.18.2
gunicorn==20.1.0
idna==2.10
importlib-metadata==3.10.0
isort==5.8.0
joblib==1.0.1
lazy-object-proxy==1.6.0
Markdown==3.3.4
mccabe==0.6.1
numpy==1.20.2
pandas==1.2.3
psycopg2==2.8.6
pylint==2.7.4
python-dateutil==2.8.1
python-whois==0.7.3
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
threadpoolctl==2.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
whitenoise==5.2.0
wrapt==1.12.1
zipp==3.4.1
